Separate signMessage and verifyMessage components
