+++ .storybook/config.js
@@ -6,9 +6,10 @@ function loadStories() {
   require('../src/components/dialog/stories');
   require('../src/components/formattedNumber/stories');
   require('../src/components/toaster/stories');
-  require('../src/components/signVerify/stories');
+  require('../src/components/signMessage/stories');
   require('../src/components/send/stories');
   require('../src/components/spinner/stories');
+  require('../src/components/verifyMessage/stories');
 }
 
 configure(loadStories, module);
