+++ src/components/header/header.js
@@ -4,8 +4,8 @@ import { IconMenu, MenuItem } from 'react-toolbox/lib/menu';
 import grid from 'flexboxgrid/dist/flexboxgrid.css';
 import logo from '../../assets/images/LISK-nano.png';
 import styles from './header.css';
-import VerifyMessage from '../signVerify/verifyMessage';
-import SignMessage from '../signVerify/signMessage';
+import VerifyMessage from '../verifyMessage';
+import SignMessage from '../signMessage';
 import RegisterDelegate from '../registerDelegate';
 import Send from '../send';
 import PrivateWrapper from '../privateWrapper';
