+++ src/components/signMessage/signMessage.js
@@ -3,7 +3,7 @@ import Input from 'react-toolbox/lib/input';
 import Lisk from 'lisk-js';
 
 import InfoParagraph from '../infoParagraph';
-import SignVerifyResult from './signVerifyResult';
+import SignVerifyResult from '../signVerifyResult';
 import ActionBar from '../actionBar';
