+++ src/components/signMessage/stories.js
@@ -2,8 +2,7 @@ import React from 'react';
 
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
-import VerifyMessage from './verifyMessage';
-import SignMessageComponent from './signMessageComponent';
+import SignMessage from './signMessage';
 
 
 const publicKey = 'c094ebee7ec0c50ebee32918655e089f6e1a604b83bcaa760293c61e0f18ab6f';
@@ -12,16 +11,11 @@ const account = {
   publicKey,
 };
 
-storiesOf('VerifyMessage', module)
-  .add('default', () => (
-    <VerifyMessage />
-  ));
-
 storiesOf('SignMessage', module)
   .add('default', () => (
-    <SignMessageComponent
+    <SignMessage
       account={account}
       closeDialog={action('closeDialog')}
       successToast={action('succesToast')}
-      />
+    />
   ));
