+++ src/components/verifyMessage/stories.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+import VerifyMessage from './index';
+
+storiesOf('VerifyMessage', module)
+  .add('default', () => (
+    <VerifyMessage />
+  ));
+
